+++ tests/test_saw.py
@@ -47,7 +47,7 @@ from scipy.sparse import csr_matrix
 
 from esa import SAW, COMError, PowerWorldError, CommandNotRespectedError, \
     Error
-from esa.saw import convert_to_windows_path
+from esa.saw import convert_to_windows_path, df_to_aux
 
 # noinspection PyUnresolvedReferences
 from tests.constants import PATH_14, PATH_14_PWD, PATH_2000, \
@@ -688,6 +688,32 @@ class GetYbusTestCase(unittest.TestCase):
         """
         self.assertIsInstance(self.saw.get_ybus(file="data/ybus.mat"),
                               csr_matrix)
+
+
+class GetAdmittanceTestCase(unittest.TestCase):
+    """Test get_branch_admittance and get_shunt_admittance function."""
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        cls.saw = SAW(PATH_14, early_bind=False)
+
+    @classmethod
+    def tearDownClass(cls) -> None:
+        # noinspection PyUnresolvedReferences
+        cls.saw.exit()
+
+    def test_get_branch_admittance_default(self):
+        """It should return two scipy csr_matrix.
+        """
+        Yf, Yt = self.saw.get_branch_admittance()
+        self.assertIsInstance(Yf, csr_matrix)
+        self.assertIsInstance(Yt, csr_matrix)
+
+    def test_get_shunt_admittance_default(self):
+        """It should return a complex numpy array.
+        """
+        Ysh = self.saw.get_shunt_admittance()
+        self.assertEqual(Ysh.dtype, np.complex_)
 
 
 class GetJacobianTestCase(unittest.TestCase):
@@ -2324,6 +2350,24 @@ class CallSimAutoTestCase(unittest.TestCase):
                        side_effect=TypeError('weird things')):
                 with self.assertRaises(TypeError):
                     saw_14.GetParametersSingleElement('bus', ['BusNum'], [1])
+
+
+class DfToAuxTestCase(unittest.TestCase):
+    """
+    Test the df_to_aux method.
+    """
+    def test_default(self):
+        df = pd.DataFrame(columns=['SOZoomHigh', 'SOZoomLow', 'SOZoomCond', 'SLName', 'SLShown',
+                                   'SOSLAnimationLayer', 'SelectableInEditMode'])
+        df.loc[len(df)] = [9900.00, 0.00, "NO ", "Borders", "YES", -1, "NO "]
+        with open("test.aux", 'w') as file:
+            df_to_aux(file, df, object_name='ScreenLayer')
+
+        with open("test.aux", 'r') as file:
+            x = len(file.readlines())
+            self.assertEqual(x, 6)
+
+        os.remove("test.aux")
 
 
 class ToNumericTestCase(unittest.TestCase):