+++ docs/rst/welcome/changelog.rst
@@ -2,6 +2,12 @@ Changes made with each ESA release are listed here. Please note that
 versions prior to 1.0.0 are not listed here, but are still available on
 `PyPi <https://pypi.org/project/esa/#history>`__.
 
+Version 1.2.6
+^^^^^^^^^^^^^
+
+* Add functions to obtain branch and shunt impedance
+* Add a helper function to convert dataframe into auxiliary(aux/axd) data section
+
 Version 1.2.5
 ^^^^^^^^^^^^^